+++ src/esgvoc/api/projects.py
@@ -10,7 +10,7 @@ from esgvoc.api._utils import (create_str_comparison_expression,
 from esgvoc.api.report import (ProjectTermError, UniverseTermError,
                                ValidationError, ValidationReport)
 from esgvoc.api.search import MatchingTerm, SearchSettings
-from esgvoc.core.db._utils import DBConnection
+from esgvoc.core.db.connection import DBConnection
 from esgvoc.core.db.models.mixins import TermKind
 from esgvoc.core.db.models.project import Collection, Project, PTerm
 from esgvoc.core.db.models.universe import UTerm
