+++ src/esgvoc/core/db/__init__.py
@@ -1,4 +1,5 @@
-from esgvoc.core.db._utils import DBConnection, read_json_file
+from esgvoc.core.db.connection import read_json_file
+from esgvoc.core.db.connection import DBConnection
 
 
 __all__ = ["DBConnection", "read_json_file"]
